# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "JFTrade sidecar API 的调试文档。Swagger UI 主要覆盖当前常用的 HTTP 调试入口，并展示实时接口的请求方式。",
        "title": "JFTrade Debug API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/execution/orders": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "execution"
                ],
                "summary": "读取执行订单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ACTIVE 表示仅活动订单",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Broker 标识",
                        "name": "brokerId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "交易环境",
                        "name": "tradingEnvironment",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "账户 ID",
                        "name": "accountId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "市场",
                        "name": "market",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/servercore.envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/servercore.executionOrdersResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/servercore.envelope"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "execution"
                ],
                "summary": "提交执行订单",
                "parameters": [
                    {
                        "description": "执行订单",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/servercore.executionPlaceOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/servercore.envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/servercore.brokerOrderCommandResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/servercore.envelope"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/servercore.envelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/servercore.envelope"
                        }
                    }
                }
            }
        },
...
        "/api/v1/execution/orders/{internalOrderId}/cancel": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "execution"
                ],
                "summary": "取消执行订单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "内部订单 ID",
                        "name": "internalOrderId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/servercore.envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/servercore.brokerOrderCommandResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/servercore.envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/servercore.envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/execution/orders/{internalOrderId}/events": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "execution"
                ],
                "summary": "读取执行订单事件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "内部订单 ID",
                        "name": "internalOrderId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/servercore.envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/servercore.executionOrderEventsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/servercore.envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "httpserver.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "servercore.brokerOrderCommandResponse": {
            "type": "object",
            "properties": {
                "accepted": {
                    "type": "boolean"
                },
                "brokerErrorCode": {
                    "type": "string"
                },
                "brokerOrderId": {
                    "type": "string"
                },
                "brokerOrderIdEx": {
                    "type": "string"
                },
                "checkedAt": {
                    "type": "string"
                },
                "internalOrderId": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "operation": {
                    "type": "string"
                },
                "orderStatus": {
                    "type": "string"
                }
            }
        },
        "servercore.envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/httpserver.APIError"
                },
                "ok": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "servercore.executionOrderEventsResponse": {
            "type": "object",
            "properties": {
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/trading.ExecutionOrderEvent"
                    }
                },
                "internalOrderId": {
                    "type": "string"
                }
            }
        },
        "servercore.executionOrdersResponse": {
            "type": "object",
            "properties": {
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/trading.ExecutionOrder"
                    }
                }
            }
        },
        "servercore.executionPlaceOrderRequest": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "string"
                },
                "brokerId": {
                    "type": "string"
                },
                "clientOrderId": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "env": {
                    "type": "string"
                },
                "market": {
                    "type": "string"
                },
                "orderType": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "number"
                },
                "remark": {
                    "type": "string"
                },
                "session": {
                    "type": "string"
                },
                "side": {
                    "type": "string"
                },
                "stopPrice": {
                    "type": "number"
                },
                "symbol": {
                    "type": "string"
                },
                "timeInForce": {
                    "type": "string"
                },
                "tradingEnvironment": {
                    "type": "string"
                }
            }
        },
...
        "trading.ExecutionOrder": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "string"
                },
                "brokerId": {
                    "type": "string"
                },
                "brokerOrderId": {
                    "type": "string"
                },
                "brokerOrderIdEx": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "filledAveragePrice": {
                    "type": "number"
                },
                "filledQuantity": {
                    "type": "number"
                },
                "internalOrderId": {
                    "type": "string"
                },
                "lastError": {
                    "type": "string"
                },
                "lastErrorCode": {
                    "type": "string"
                },
                "lastErrorSource": {
                    "type": "string"
                },
                "market": {
                    "type": "string"
                },
                "orderType": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "requestedPrice": {
                    "type": "number"
                },
                "requestedQuantity": {
                    "type": "number"
                },
                "side": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "sourceDetail": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "submittedAt": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "tradingEnvironment": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "trading.ExecutionOrderEvent": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "eventType": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "internalOrderId": {
                    "type": "string"
                },
                "nextStatus": {
                    "type": "string"
                },
                "payloadJson": {
                    "type": "string"
                },
                "previousStatus": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AdministratorBearer": {
            "description": "Bearer \u003cJFTrade administrator key\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "AdministratorSession": {
            "type": "apiKey",
            "name": "jftrade_admin_session",
            "in": "cookie"
        }
    }
}